+++ Packs/Legacy/Integrations/AwakeSecurity/AwakeSecurity.py
@@ -20,12 +20,40 @@ identifier = credentials["identifier"]
 password = credentials["password"]
 suspicious_threshold = params["suspicious_threshold"]
 malicious_threshold = params["malicious_threshold"]
-authTokenRequest = {
-    "loginUsername": identifier,
-    "loginPassword": password
-}
-authTokenResponse = requests.post(prefix + "/authtoken", json=authTokenRequest, verify=verify)
-authToken = authTokenResponse.json()["token"]["value"]
+
+
+def get_authtoken():
+    headers = {
+        "Content-Type": "application/json"
+    }
+    authTokenRequest_v1 = {
+        "loginUsername": identifier,
+        "loginPassword": password
+    }
+
+    authTokenRequest_v2 = {
+        "username": {
+            "username": identifier
+        },
+        "plaintextPassword": {
+            "value": password
+        }
+    }
+    authTokenResponse = requests.post(prefix + "/authtoken", json=authTokenRequest_v1, verify=verify, headers=headers)
+    if authTokenResponse.status_code == 200:
+        return authTokenResponse.json()["token"]["value"]
+
+    else:
+        authTokenResponse = requests.post(prefix + "/authtoken", json=authTokenRequest_v2, verify=verify,
+                                          headers=headers)
+        if authTokenResponse.status_code == 200:
+            return authTokenResponse.json()["token"]["value"]
+
+        else:
+            raise DemistoException("Authtoken could not be fetched - check the given credentials.")
+
+
+authToken = get_authtoken()
 headers = {
     "Authentication": ("access " + authToken),
     'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, '
@@ -90,7 +118,7 @@ def returnResults(contents, outerKey, innerKey, humanReadable, dbotScore):
         "AwakeSecurity": contents,
     }
     entryContext = {
-        ("AwakeSecurity." + outerKey + "(val." + innerKey + "===obj." + innerKey + ")"): contents,
+        ("AwakeSecurity." + outerKey + "(val." + innerKey + "== obj." + innerKey + ")"): contents,
     }
     if dbotScore is not None:
         machineReadable["DBotScore"] = dbotScore
@@ -307,7 +335,7 @@ def queryActivities():
     humanReadable = displayTable(contents, humanReadableFields)
     for content in contents:
         content["query"] = q
-    returnResults(contents, "Activities", "query", humanReadable, None)
+    returnResults(contents, "Activities", "activityId", humanReadable, None)
 
 
 def queryDevices():
@@ -328,7 +356,7 @@ def queryDevices():
     humanReadable = displayTable(contents, humanReadableFields)
     for content in contents:
         content["query"] = q
-    returnResults(contents, "Devices", "query", humanReadable, None)
+    returnResults(contents, "Devices", "deviceId", humanReadable, None)
 
 
 def queryDomains():
@@ -350,7 +378,7 @@ def queryDomains():
     humanReadable = displayTable(contents, humanReadableFields)
     for content in contents:
         content["query"] = q
-    returnResults(contents, "Domains", "query", humanReadable, None)
+    returnResults(contents, "Domains", "name", humanReadable, None)
 
 
 def pcapDownload():
@@ -463,7 +491,7 @@ try:
     elif command == "device":
         lookupDevice()
 
-except Exception, e:
+except Exception as e:
     if command == "fetch-incidents":
         raise
     LOG(e)